Refactor of Access Boston authentication
Adds support for a second "forgot password" SAML app to handle the
authentication of forgot password requests.

Adds stand-in for Forgot Password form, but it’s not hooked up to any
functionality yet.

Factors out the Next.js route helper so that we can add separate
authentication for the "/forgot" route from the rest of the Next app’s
authentication.