Makes production deploys use the Slack bot
Adds code to call the GitHub API to move the branch to the new commit
before triggering CodeBuild.

Updates .travis.yml to have the notifications to the bot be the deploy
part of the test stage to avoid needing to re-install node dependencies
(and since develop is now the only branch Travis cares about).