Make indexer sensitive to meta failures
This throws an error when a meta message is a failure, such as when
Salesforce errors during a long-poll reconnect. Previously these were
ignored and we just stopped receiving messages.

Updates the error handling to complete the RxJS stream rather than
immediately quit, which allows existing processing to cleanly complete.

Also updates to latest bugfix version of Comet 3.1 line and adds http
proxy support in more places for better debugging with Charles.