Adds simple cache support to fetchGraphql
Caching is done by passing in a cache object when creating the
fetchGraphql function instance, and providing a cache key any time you
want to potentially read from / write to the cache.

Makes it straightforward to cache requests made during server-side
rendering so that they’re not re-fetched in the browser.

Fixes #188