Consistently treats dates as midnight UTC
The contract for the API is now to use midnight UTC, and we render the
date in emails in UTC as well.

Prevents problems where birth dates were being given a time early in the
morning in UTC that was the day before when we rendered the time out in
eastern.

Fixes #368