Upgrades to Emotion 10
We’re using the old-style "css" function, so no need to add the Babel
plugin right now to support the "css" prop in JSX.

Fixes Emotion hydration by putting it in the raw file rather than App
constructor. This prevents double-insertion of styles.

Also fixes the public notices CSS that was pushing the bottom bar off
the screen.